Netscape Navigator 9.0.6
This is a copy of the original Netscape Navigator 9.0.6 (for windows). Unfortunately, this happens also to be the last release of Netscape Navigator. 

AOL (America Online) owns the rights and trademarks. 
I include it only as a reference point for inspiration when developing any styles, themes, or other modifications designed for Mozilla Firefox. 
It comes with no warranty or guarantee of functionality. AOL (America Online) and its respected partners retain all rights.